upgrade code/ merge win&lose sys/ fix button goToMenu
