+++ Assets/Prefabs/Towers/TowerImgPrefab.prefab
@@ -27,7 +27,7 @@ Transform:
   m_GameObject: {fileID: 6478238751360883004}
   serializedVersion: 2
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 0.111821115, y: 5.886581, z: -9}
+  m_LocalPosition: {x: -0.44728446, y: -0.9504792, z: -9}
   m_LocalScale: {x: 0.3, y: 0.3, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children: []
