+++ Assets/Scripts/Authoring/CastleAuthoring.cs
@@ -16,10 +16,10 @@ class CastleAuthoringBaker : Baker<CastleAuthoring>
         AddComponent<CastleComponent>(castleEntity);
         AddComponent<OffSceneComponent>(castleEntity);
 
-        DynamicBuffer<WayPointsComponent> path = AddBuffer<WayPointsComponent>(castleEntity);
+        DynamicBuffer<WayPointsBufferElements> path = AddBuffer<WayPointsBufferElements>(castleEntity);
         foreach (var point in authoring.Path)
         {
-            WayPointsComponent wp = default;
+            WayPointsBufferElements wp = default;
             wp.Value = point;
             path.Add(wp);
         }
