--- Assets/Scripts/BufferElements/BurningBufferElement.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 713d6dc282cdada4cbcae4c375269f1d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
