--- Assets/Scripts/BufferElements/WayPointsComponent.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 4c58583502588464c94965757dbc07a6
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
