+++ Assets/Scripts/Components/BufferElements/BurningBufferElement.cs
