+++ Assets/Scripts/Components/BufferElements/DamageBufferElement.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a0d565fe282dc2e42b2a0bef58f88742
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
