--- Assets/Scripts/MonoBehaviour/UI/BackToMenu.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 85e2936c1a6a41bb8400aa49d897f80b
-timeCreated: 1701552422
