+++ Assets/Scripts/Systems/Damage/BurningSystem.cs
