+++ Assets/Scripts/Systems/Damage/BurningSystem.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e52ece3244526fd4b80a3034f0565f45
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
