+++ Assets/Scripts/Systems/Damage/EffectResolverSystem.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 30c6b32236e637d459f2c3f61f7b4130
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
