--- Assets/Scripts/Systems/DamageSystem.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 19bcb79c5972c6f4999eda80a47fd3a8
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
