--- Assets/Scripts/Systems/LoseLevelSystem.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 07ee717d952085b4c9c8d05729f41a5b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
