--- Assets/Scripts/Systems/MoveBulletSystem.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 8597d459ac8ee7f4d80e5127ad50285a
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
