+++ Assets/Scripts/Systems/MoveSpawnDestroy/DestroySystem.cs
