--- Assets/Scripts/Systems/SpawnBulletSystem.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: ba8565cd2ae74c82ac6b93eacc8877d3
-timeCreated: 1674833466
